Adding README file and Instructions
While most of the comments in the source code for this project generally explains how to use the project, I have missed the direct log-in information that can be found in the Users table of the "climateChangeDatabase.db" file. In addition to adding to the formatting of the project, I hope that this README file adds some further readability for the project as a whole and acts as a proper introduction to other users of this project!